functions-cm-automation/requirements.txt
db-dtypes==1.2.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
pandas==2.2.2
pyarrow==16.1.0
urllib3==2.2.1
wcwidth==0.2.13